add: offer - cached
