+++ app/tests/offer/test_offer.py
@@ -10,7 +10,7 @@ from sqlmodel.ext.asyncio.session import AsyncSession
 from tests.utils import url_reverse
 from tests.conftest import engine
 from entities.offer.models_offer import OfferLinkModel, OfferLinkInSchema
-from entities.offer.routers_offer import OfferFilter
+from entities.offer.choices_offer import OfferFilter
 from entities.offer.schemas_offer import OfferLinkOptionalSchema
