+++ app/tests/offer/test_offer_full.py
@@ -4,7 +4,7 @@ from fastapi import status
 from sqlmodel.ext.asyncio.session import AsyncSession
 
 from tests.utils import url_reverse
-from entities.offer.routers_offer import OfferFilter
+from entities.offer.choices_offer import OfferFilter
 from entities.offer.schemas_offer import OfferFullResponseSchema
